+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+# 3.7.0
+- Make entities in the air be affect by the movement speed attribute
+- Reduce the resolution of the mod icon. Why was it that big
+- Glass armor
+
 # 3.6.0
 - Buffer shulker bullets a fuck ton
 - Add "Splash" Trait Modifier to Tinker's Construct